:recycle: refactor: modularize MapView.svelte
Completed all phases of MapView refactoring. Extracted interaction management, canvas rendering, and DOM overlays into separate focused modules. Reduced MapView.svelte from 1500+ lines to ~330 lines. Resolves #789.
